Convert module to TypeScript
